fix int size_t comparison in scoreboard

+++ game/scoreboard.cpp
@@ -186,7 +186,7 @@ namespace game
         }
         LOOP_START(id, stack);
         std::vector<gameent *> &p = *team < 0 ? spectators : teamplayers[*team];
-        for(int i = 0; i < p.size(); i++)
+        for(size_t i = 0; i < p.size(); i++)
         {
             loopiter(id, stack, p[i]->clientnum);
             execute(body);
